Remove predefines that are breaking the doxygen build.
I have no idea why they are, but the macros appear to be defined in the
appropriate header, so maybe they're not needed anyway?

Btw. to replicate the failure you need to delete the temporary files
under bin.v2, as boost.build doesn't detect the failure, and continue
with the old copies from the last successful build. I'm using doxygen
1.8.8, maybe this is fixed in later versions?