Fixed small_vector default constructor noexcept specification.
The default constructor used to incorrectly apply
is_nothrow_default_constructible trait to the template argument of
the small_vector class template, which is void by default. The fix is to
apply the trait to the actual allocator type.

Next is the fix for noexcept specifications of the small_vector_allocator
class members, including the default constructor, which was previously not
declared and a non-noexcept initializing constructor was used in its stead.
The fix is to explicitly specify the default constructor with a proper
noexcept specification. While at it, other constructors and assignment
operators were also fixed with proper noexcept specifications.

As a result, small_vector default constructor noexcept specification now
properly reflects whether the default constructor of the allocator is noexcept.